add arm demo
